3 iteration

+++ event-manager/src/main/java/dev/sorokin/eventmanager/service/LocationService.java
@@ -1,7 +1,9 @@
 package dev.sorokin.eventmanager.service;
 
+import dev.sorokin.eventmanager.domain.EventStatus;
 import dev.sorokin.eventmanager.domain.Location;
 import dev.sorokin.eventmanager.dto.LocationDto;
+import dev.sorokin.eventmanager.entity.LocationEntity;
 import dev.sorokin.eventmanager.mapper.LocationMapper;
 import dev.sorokin.eventmanager.repository.LocationRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -15,13 +17,16 @@ public class LocationService {
 
     private final LocationMapper locationMapper;
     private final LocationRepository locationRepository;
+    private final EventService eventService;
 
     public LocationService(
             LocationMapper locationMapper,
-            LocationRepository locationRepository
+            LocationRepository locationRepository,
+            EventService eventService
     ) {
         this.locationMapper = locationMapper;
         this.locationRepository = locationRepository;
+        this.eventService = eventService;
     }
 
 
@@ -51,19 +56,18 @@ public class LocationService {
                     .formatted(id));
         }
 
+        var events = eventService.getAllEventsByLocationId(id);
+
+        boolean hasActiveEvents = events.stream()
+                        .anyMatch(event -> event.getStatus() == EventStatus.WAIT_START ||
+                                event.getStatus() == EventStatus.STARTED);
+
+        if(hasActiveEvents) {
+            throw new IllegalArgumentException("Cannot delete location with active events");
+        }
+
         locationRepository.deleteById(id);
     }
-
-
-    public Location findLocationById(Long id) {
-        return locationMapper.toDomain(
-                locationRepository.findById(id)
-                        .orElseThrow(() -> new EntityNotFoundException(
-                                "Location with id=%s not found".formatted(id)
-                        ))
-        );
-    }
-
 
     public Location updateLocation(
             Long id,
@@ -85,6 +89,11 @@ public class LocationService {
                     .formatted(newName));
         }
 
+        if(!checkCapacityForUpdate(id, locationDtoToUpdate.capacity(), locationEntityToUpdate)) {
+            throw new IllegalArgumentException("Cannot decrease location capacity below maxPlaces of existing events");
+        }
+
+
         locationEntityToUpdate.setName(locationToUpdate.name());
         locationEntityToUpdate.setAddress(locationToUpdate.address());
         locationEntityToUpdate.setCapacity(locationToUpdate.capacity());
@@ -93,5 +102,42 @@ public class LocationService {
         return locationMapper.toDomain(
                 locationRepository.save(locationEntityToUpdate)
         );
+    }
+
+    public Location findLocationById(Long id) {
+        return locationMapper.toDomain(
+                locationRepository.findById(id)
+                        .orElseThrow(() -> new EntityNotFoundException(
+                                "Location with id=%s not found".formatted(id)
+                        ))
+        );
+    }
+
+    private boolean checkCapacityForUpdate(
+            Long id,
+            Integer newCapacity,
+            LocationEntity oldLocation
+    ) {
+        if (newCapacity > oldLocation.getCapacity()) {
+            return true;
+        }
+
+        if (Objects.equals(newCapacity, oldLocation.getCapacity())) {
+            return true;
+        }
+
+        var events = eventService.getAllEventsByLocationId(id);
+
+        boolean hasActiveEventWithMorePlaces = events.stream()
+                .filter(event -> event.getStatus() == EventStatus.WAIT_START ||
+                        event.getStatus() == EventStatus.STARTED)
+                .anyMatch(event -> event.getMaxPlaces() > newCapacity);
+
+        return !hasActiveEventWithMorePlaces;
+    }
+
+    public LocationEntity findLocationEntityById(Long id) {
+        return locationRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Location not found"));
     }
 }
